Add a utility arrangement for local serving of files and cgi functionality.
Run ./serve.py to start a localhost:8000 simple cgi server on this directory.
(Run ./serve.py 8888 foo.html to use port 8888 and start at page foo.html,
defaults are port 8000 and page index.html).  ./serve.py also runs a browser
on the given HTML page (default: index.html).  cgi-bin is a symbolic link
to the current directory (to request CGI service); as a further convenience,
serve.py, when executed as cgi-bin/serve.py, runs as CGI (by default serving
a plain-text "hello world" page;-).

The idea (next TODO) is to have cgi-bin/serve.py?png=...&tc serve the
requested png file AND log the rest of the parameters for future reference;
I plan to add this functionality in the very next commit (together with a
mod to the crosshair.html page to make use of it;-).